refactor(db): change db package structure for better readability
