+++ db/sqlc.yml
@@ -5,8 +5,8 @@ sql:
     queries: "./queries"
     gen:
       go:
-        package: "db"
-        out: "./db"
+        package: "sqlc"
+        out: "./sqlc"
         sql_package: "pgx/v5"
         emit_json_tags: true
         emit_prepared_queries: false
