+++ db/sqlc/convertors.go
@@ -1,4 +1,4 @@
-package db
+package sqlc
 
 import (
 	"github.com/google/uuid"
